Move fake data generation to separate file
